Add transitive dependencies to the AUTODEP mechanism
Any erlang application may depend on an arbitrary set of other applications,
which in turn may have their own requirements. This places the impetus on
the maintainer of the appropriate .app/.rel files to correctly state all
explicit and transitive dependencies.

In the ideal world, all of an applications dependencies would be managed
as erlang-[otp|std] maven coordinates generated by the maven-erlang-plugin,
in which case transitive dependencies would be naturally handled by maven.
However, realistically not all applications will be packaged in maven
(OTP apps, for instance).  Therefore, it would be nice if we can still
support transitive dependencies using the information inherent to
native erlang.  Fortunately, erlang already has this metadata neatly
packaged up for us in in the {application, $list} tuple for each .app.

This patch exploits this information by taking explicit dependencies
(declared as standard coordinate deps, and/or the additionalAutoDeps parameter)
and walks backwards through the data.  This is then used to generate
a complete list of required applications, explcit and transitive, that we
use as the basis of the ${AUTODEPS} substitution in the .rel

Signed-off-by: Gregory Haskins <ghaskins@novell.com>