Add "systemDependencies" parameter to .app packager
We plan on exploiting the (presumably) accurate application dependency
information embedded in .app files later in the series.  Therefore
we need to ensure that we formulate the .app files correctly.  Given
the trend to have maven-erlang-plugin automatically manage the generation
of our .app/.rel metadata, we need to provide an easy way to augment
the ${APPLICATION} with our true dependencies.

This information would ideally come from pure maven dependencies, but
not all artifacts will have been built from maven.  Some dependencies
can be assumed to already be present on the build system (for instance,
OTP apps like SASL and MNESIA).  One might be inclined to use the
<scope>system</scope> dependency, but this doesn't work quite as expected
since it assumes the target coordinate is in some package format that
one or more maven plugins knows how to handle.  We could then write
an ArchiveHandler for say "erlang-installed" to hande system-scoped
artifacts, but this leads to a bunch of ugly side-effects, such as the
notion that we need to now declare a complete coordinate tuple (group,
artifact, version, type, scope, and possibly, classifier) when all
we really want is the artifactId.  The rest can be resolved by the
REL packager later.

Therefore, this patch introduces a new parameter "systemDependencies"
which takes a comma separated list of "system" artifacts (read: erlang
applications installed somewhere in the ERL_LIBS path).  These declared
items will be emitted to the ${APPLICATION} variable in the .app, when
applicable.

Example POM usage: <systemDependencies>sasl, mnesia</systemDependencies>

Signed-off-by: Gregory Haskins <ghaskins@novell.com>